# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Reserva de turnos odontológicos",
        "title": "Api de turnos Odontologicos",
        "termsOfService": "https://developers.ctd.com.ar/es_ar/terminos-y-condiciones",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/odontologos": {
            "post": {
                "description": "Crea un nuevo odontólogo con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontólogos"
                ],
                "summary": "Crear un nuevo odontólogo",
                "parameters": [
                    {
                        "description": "Datos del odontólogo a crear",
                        "name": "odontologo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Odontólogo creado correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/odontologos/{idOdontologo}": {
            "get": {
                "description": "Obtiene un odontólogo por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontólogos"
                ],
                "summary": "Obtener un odontólogo por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del odontólogo a obtener",
                        "name": "idOdontologo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Odontólogo obtenido correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un odontólogo por su ID con todos los campos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontólogos"
                ],
                "summary": "Actualizar un odontólogo por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del odontólogo a actualizar",
                        "name": "idOdontologo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del odontólogo a actualizar",
                        "name": "odontologo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Odontólogo actualizado correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Error al actualizar el odontólogo",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un odontólogo por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontólogos"
                ],
                "summary": "Eliminar un odontólogo por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del odontólogo a eliminar",
                        "name": "idOdontologo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Odontólogo eliminado correctamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente un odontólogo por su ID con campos específicos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontólogos"
                ],
                "summary": "Actualizar parcialmente un odontólogo por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del odontólogo a actualizar",
                        "name": "idOdontologo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar",
                        "name": "odontologo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Odontólogo actualizado correctamente",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    },
                    "400": {
                        "description": "Error de solicitud",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No se encontró el ID indicado",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Error al actualizar el odontólogo",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Odontologo": {
            "type": "object",
            "required": [
                "apellidoOdontologo",
                "matriculaOdontologo",
                "nombreOdontologo"
            ],
            "properties": {
                "apellidoOdontologo": {
                    "type": "string"
                },
                "idOdontologo": {
                    "type": "integer"
                },
                "matriculaOdontologo": {
                    "type": "string"
                },
                "nombreOdontologo": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.Request": {
            "type": "object",
            "properties": {
                "apellidoOdontologo": {
                    "type": "string"
                },
                "matriculaOdontologo": {
                    "type": "string"
                },
                "nombreOdontologo": {
                    "type": "string"
                }
            }
        }
    }
}